Refactor chunk definition
