+++ src/client/terrain/util/mod.rs
@@ -2,10 +2,14 @@ pub mod blocks;
 pub mod buffer_serializer;
 pub mod chunk;
 pub mod generator;
+pub mod chunk;
+pub mod generator;
 pub mod mesher;
 
 pub use blocks::*;
 pub use buffer_serializer::*;
+pub use chunk::*;
+pub use generator::*;
 pub use chunk::*;
 pub use generator::*;
 pub use mesher::*;
