+++ src/server/terrain/util/mod.rs
@@ -1,7 +1,5 @@
 pub mod blocks;
-pub mod chunk;
 pub mod generator;
 
 pub use blocks::*;
-pub use chunk::*;
 pub use generator::*;
